+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog
+
+## 6.3.0-wip
+
+- Fixes `hide-isolated` filter was not being applied to the `print` and
+  `print-mermaid` actions.
+
 ## 6.2.0+1
 
 - Documentation updates.